@@ -1,47 +1,17 @@
 {
     "name": "practitioner-app",
-    "version": "0.1.0",
+    "version": "1.0.0",
     "private": true,
+    "type": "module",
     "scripts": {
-        "start": "concurrently \"PORT=3010 react-scripts start\" \"node server.js\"",
-        "build": "react-scripts build",
-        "test": "react-scripts test",
-        "eject": "react-scripts eject",
+        "start": "concurrently \"PORT=3010 vite\" \"node server.js\"",
+        "dev": "vite",
+        "build": "vite build",
+        "preview": "vite preview",
+        "test": "vitest",
         "server": "node server.js",
-        "heroku-start": "node --max-old-space-size=512 node_modules/.bin/react-scripts start"
-    },
-    "dependencies": {
-        "@emotion/react": "^11.11.1",
-        "@emotion/styled": "^11.11.0",
-        "@mui/icons-material": "^5.14.16",
-        "@mui/lab": "^5.0.0-alpha.152",
-        "@mui/material": "^5.14.17",
-        "@rjsf/core": "^5.13.6",
-        "@rjsf/mui": "^5.13.6",
-        "@rjsf/utils": "^5.13.6",
-        "@rjsf/validator-ajv8": "^5.13.6",
-        "@testing-library/jest-dom": "^5.17.0",
-        "@testing-library/react": "^13.4.0",
-        "@testing-library/user-event": "^13.5.0",
-        "@types/fhir": "^0.0.35",
-        "@types/jest": "^27.5.2",
-        "@types/node": "^16.18.60",
-        "@types/react": "^18.2.36",
-        "@types/react-dom": "^18.2.14",
-        "cors": "^2.8.5",
-        "date-fns": "^2.30.0",
-        "express": "^4.21.2",
-        "fhir-questionnaire-json-schema": "github:HTD-Health/fhir-questionnaire-json-schema",
-        "fhirclient": "^2.5.2",
-        "jsonschema": "^1.4.1",
-        "morgan": "^1.10.0",
-        "react": "^18.2.0",
-        "react-dom": "^18.2.0",
-        "react-query": "^3.39.3",
-        "react-router-dom": "^6.18.0",
-        "react-scripts": "5.0.1",
-        "typescript": "^4.9.5",
-        "web-vitals": "^2.1.4"
+        "lint": "eslint .",
+        "format": "prettier --write ."
     },
     "engines": {
         "node": "22.x"
@@ -65,9 +35,43 @@
         ]
     },
     "devDependencies": {
-        "@kensho-technologies/eslint-config": "^24.1.0",
-        "concurrently": "^8.2.2",
-        "eslint": "^8.29.0",
-        "ts-loader": "^9.4.2"
+        "@eslint/eslintrc": "^3.3.1",
+        "@eslint/js": "^9.26.0",
+        "@tanstack/eslint-plugin-query": "^5.74.7",
+        "@types/fhir": "^0.0.35",
+        "@types/react": "^19.1.3",
+        "@types/react-dom": "^19.1.3",
+        "@typescript-eslint/parser": "^8.32.1",
+        "@vitejs/plugin-react": "^4.4.1",
+        "concurrently": "^9.1.2",
+        "eslint": "^9.26.0",
+        "eslint-import-resolver-alias": "^1.1.2",
+        "eslint-plugin-import": "^2.31.0",
+        "eslint-plugin-react": "^7.37.5",
+        "eslint-plugin-react-hooks": "^5.2.0",
+        "eslint-plugin-unused-imports": "^4.1.4",
+        "prettier": "^3.5.3",
+        "typescript": "^5.8.3",
+        "typescript-eslint": "^8.32.1",
+        "vite": "^6.3.5",
+        "vitest": "^3.1.3"
+    },
+    "dependencies": {
+        "@emotion/react": "^11.14.0",
+        "@emotion/styled": "^11.14.0",
+        "@mui/icons-material": "^6.4.11",
+        "@mui/material": "^6.4.11",
+        "@rjsf/core": "^5.24.10",
+        "@rjsf/mui": "^5.24.10",
+        "@rjsf/utils": "^5.24.10",
+        "@rjsf/validator-ajv8": "^5.24.10",
+        "@tanstack/react-query": "^5.75.5",
+        "date-fns": "^4.1.0",
+        "fhir-questionnaire-json-schema": "github:HTD-Health/fhir-questionnaire-json-schema",
+        "fhirclient": "^2.5.5",
+        "morgan": "^1.10.0",
+        "react": "^19.1.0",
+        "react-dom": "^19.1.0",
+        "react-router-dom": "^7.5.3"
     }
 }
