@@ -1,6 +1,6 @@
-import { QueryClient, QueryClientProvider } from 'react-query';
-
-import AppRouter from 'app_router';
+import AppRouter from '@/app_router';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { JSX } from 'react';
 
 const App = (): JSX.Element => {
     const queryClient = new QueryClient();
