@@ -1,13 +1,14 @@
 import { BrowserRouter as HashRouter, Navigate, Route, Routes } from 'react-router-dom';
 
-import AssignedFormsContainer from 'pages/assigned_forms/assigned_forms_container';
-import FilledFormsContainer from 'pages/filled_forms/filled_forms_container';
-import FormFill from 'pages/form_fill/form_fill';
-import FormsContainer from 'pages/forms/forms_container';
-import GoalsListPage from 'pages/goals/goals_list_page';
-import PatientProfile from 'pages/patient_profile/patient_profile';
-import ResponseView from 'pages/response_view/response_view';
-import routes from 'routes';
+import AssignedFormsContainer from '@/pages/assigned_forms/assigned_forms_container';
+import FilledFormsContainer from '@/pages/filled_forms/filled_forms_container';
+import FormFill from '@/pages/form_fill/form_fill';
+import FormsContainer from '@/pages/forms/forms_container';
+import GoalsListPage from '@/pages/goals/goals_list_page';
+import PatientProfile from '@/pages/patient_profile/patient_profile';
+import ResponseView from '@/pages/response_view/response_view';
+import routes from '@/routes';
+import { JSX } from 'react';
 
 const AppRouter = (): JSX.Element => (
     <HashRouter>
