@@ -1,26 +1,27 @@
-import { Box, Card, CircularProgress, Typography } from '@mui/material';
 import WarningIcon from '@mui/icons-material/Warning';
-import { format } from 'date-fns';
-import type { AllergyIntolerance } from 'fhir/r4';
+import { Box, Card, CircularProgress, Typography } from '@mui/material';
+import type { AllergyIntolerance, BundleEntry } from 'fhir/r4';
 
-import { useGetAllergies } from 'api/queries';
+import { useGetAllergies } from '@/api/queries';
+import { format } from 'date-fns';
+import { JSX } from 'react';
 
 const AllergiesSummary = (): JSX.Element => {
     const { data, isLoading } = useGetAllergies();
 
     const renderAllergy = (allergy: AllergyIntolerance): JSX.Element => {
-        const recordedDate = allergy.recordedDate 
+        const recordedDate = allergy.recordedDate
             ? format(new Date(allergy.recordedDate), 'MMM d, yyyy')
             : 'Date not recorded';
 
-        const criticality = allergy.criticality 
+        const criticality = allergy.criticality
             ? allergy.criticality.charAt(0).toUpperCase() + allergy.criticality.slice(1).toLowerCase()
             : 'Unknown';
 
         return (
-            <Box 
+            <Box
                 key={allergy.id}
-                sx={{ 
+                sx={{
                     display: 'flex',
                     alignItems: 'center',
                     gap: '0.5rem',
@@ -32,18 +33,18 @@ const AllergiesSummary = (): JSX.Element => {
             >
                 <WarningIcon sx={{ color: 'warning.main' }} />
                 <Box>
-                    <Typography 
+                    <Typography
                         variant="body1"
-                        sx={{ 
+                        sx={{
                             fontWeight: 500,
                             color: 'text.primary'
                         }}
                     >
                         {allergy.code?.text || 'Unnamed allergy'}
                     </Typography>
-                    <Typography 
+                    <Typography
                         variant="body2"
-                        sx={{ 
+                        sx={{
                             color: 'text.secondary',
                             fontSize: '0.875rem'
                         }}
@@ -68,11 +69,7 @@ const AllergiesSummary = (): JSX.Element => {
             );
         }
 
-        return (
-            <>
-                {data.entry.map((entry) => renderAllergy(entry.resource as AllergyIntolerance))}
-            </>
-        );
+        return <>{data.entry.map((entry: BundleEntry) => renderAllergy(entry.resource as AllergyIntolerance))}</>;
     };
 
     return (
@@ -83,13 +80,13 @@ const AllergiesSummary = (): JSX.Element => {
                 transition: 'all 0.2s ease-in-out',
                 '&:hover': {
                     transform: 'translateY(-2px)',
-                    boxShadow: '0 4px 6px rgba(0,0,0,0.1), 0 2px 4px rgba(0,0,0,0.06)',
+                    boxShadow: '0 4px 6px rgba(0,0,0,0.1), 0 2px 4px rgba(0,0,0,0.06)'
                 }
             }}
         >
-            <Typography 
-                variant="h6" 
-                sx={{ 
+            <Typography
+                variant="h6"
+                sx={{
                     mb: '1rem',
                     fontWeight: 600,
                     color: 'text.primary'
@@ -102,4 +99,4 @@ const AllergiesSummary = (): JSX.Element => {
     );
 };
 
-export default AllergiesSummary; 
+export default AllergiesSummary;
