@@ -1,8 +1,9 @@
 import ArrowRightAltIcon from '@mui/icons-material/ArrowRightAlt';
-import { Box, Button, Card, Typography, Tooltip } from '@mui/material';
-import { format } from 'date-fns';
-import { useNavigate } from 'react-router-dom';
 import DescriptionIcon from '@mui/icons-material/Description';
+import { Box, Button, Card, Tooltip, Typography } from '@mui/material';
+import { format } from 'date-fns';
+import { JSX } from 'react';
+import { useNavigate } from 'react-router-dom';
 
 type FilledFormItemProps = {
     name: string;
@@ -25,29 +26,25 @@ const FilledFormItem = (props: FilledFormItemProps): JSX.Element => {
                 transition: 'all 0.2s ease-in-out',
                 '&:hover': {
                     transform: 'translateY(-2px)',
-                    boxShadow: '0 4px 6px rgba(0,0,0,0.1), 0 2px 4px rgba(0,0,0,0.06)',
+                    boxShadow: '0 4px 6px rgba(0,0,0,0.1), 0 2px 4px rgba(0,0,0,0.06)'
                 }
             }}
         >
             <Box sx={{ display: 'flex', alignItems: 'center', gap: '1rem' }}>
                 <DescriptionIcon sx={{ color: 'primary.main' }} />
                 <Box>
-                    <Typography 
-                        variant="h6" 
+                    <Typography
+                        variant="h6"
                         color="text.primary"
                         sx={{
                             fontWeight: 500,
-                            letterSpacing: '-0.01em',
+                            letterSpacing: '-0.01em'
                         }}
                     >
                         {name}
                     </Typography>
                     {date && (
-                        <Typography 
-                            variant="body2" 
-                            color="text.secondary"
-                            sx={{ mt: '0.25rem' }}
-                        >
+                        <Typography variant="body2" color="text.secondary" sx={{ mt: '0.25rem' }}>
                             Completed on: {format(new Date(date), 'iii, MM/dd/yyyy HH:mm:ss')}
                         </Typography>
                     )}
@@ -67,8 +64,8 @@ const FilledFormItem = (props: FilledFormItemProps): JSX.Element => {
                             '&:hover': {
                                 borderColor: 'primary.dark',
                                 backgroundColor: 'primary.main',
-                                color: 'white',
-                            },
+                                color: 'white'
+                            }
                         }}
                     >
                         View
