@@ -1,15 +1,14 @@
-import PersonIcon from '@mui/icons-material/Person';
 import ArrowCircleLeftIcon from '@mui/icons-material/ArrowCircleLeft';
-import AppBar from '@mui/material/AppBar';
+import PersonIcon from '@mui/icons-material/Person';
 import { Box, IconButton, Tooltip } from '@mui/material';
+import AppBar from '@mui/material/AppBar';
 import Typography from '@mui/material/Typography';
 import type { Practitioner } from 'fhir/r4';
-import { useEffect, useState } from 'react';
+import { JSX, useEffect, useState } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
 
-
-import { useGetUser } from 'api/queries';
-import CustomSnackbar from 'components/custom_snackbar/custom_snackbar';
+import { useGetUser } from '@/api/queries';
+import CustomSnackbar from '@/components/custom_snackbar/custom_snackbar';
 
 const SmartAppBar = (): JSX.Element => {
     const [errorSnackbar, setErrorSnackbar] = useState(false);
@@ -43,16 +42,16 @@ const SmartAppBar = (): JSX.Element => {
             return <Box />;
         }
         return (
-            <Box 
-                sx={{ 
-                    display: 'flex', 
+            <Box
+                sx={{
+                    display: 'flex',
                     alignItems: 'center',
                     backgroundColor: 'rgba(255, 255, 255, 0.1)',
                     borderRadius: '8px',
                     padding: '8px 16px',
                     transition: 'all 0.2s ease-in-out',
                     '&:hover': {
-                        backgroundColor: 'rgba(255, 255, 255, 0.2)',
+                        backgroundColor: 'rgba(255, 255, 255, 0.2)'
                     }
                 }}
             >
@@ -65,14 +64,14 @@ const SmartAppBar = (): JSX.Element => {
     };
 
     return (
-        <AppBar 
+        <AppBar
             position="relative"
             elevation={0}
             sx={{
                 borderBottom: '1px solid',
                 borderColor: 'grey.200',
                 bgcolor: 'primary.main',
-                color: 'white',
+                color: 'white'
             }}
         >
             <Box
@@ -85,7 +84,7 @@ const SmartAppBar = (): JSX.Element => {
                     maxWidth: '1400px',
                     margin: '0 auto',
                     width: '100%',
-                    position: 'relative',
+                    position: 'relative'
                 }}
             >
                 <Box sx={{ display: 'flex', alignItems: 'center', pl: '2rem' }}>
@@ -93,13 +92,13 @@ const SmartAppBar = (): JSX.Element => {
                         <Tooltip title="Go back">
                             <IconButton
                                 onClick={() => navigate(-1)}
-                                sx={{ 
+                                sx={{
                                     color: 'white',
                                     transition: 'all 0.2s ease-in-out',
                                     mr: '1rem',
                                     '&:hover': {
                                         transform: 'translateX(-2px)',
-                                        color: 'grey.100',
+                                        color: 'grey.100'
                                     }
                                 }}
                             >
@@ -108,19 +107,19 @@ const SmartAppBar = (): JSX.Element => {
                         </Tooltip>
                     )}
                     <Box sx={{ display: 'flex', alignItems: 'center' }}>
-                        <Typography 
-                            variant="h5" 
-                            sx={{ 
+                        <Typography
+                            variant="h5"
+                            sx={{
                                 fontWeight: 700,
                                 letterSpacing: '0.02em',
                                 color: 'white',
-                                lineHeight: 1,
+                                lineHeight: 1
                             }}
                         >
                             HTD
                         </Typography>
-                        <Box 
-                            sx={{ 
+                        <Box
+                            sx={{
                                 width: '6px',
                                 height: '6px',
                                 borderRadius: '50%',
@@ -131,9 +130,7 @@ const SmartAppBar = (): JSX.Element => {
                         />
                     </Box>
                 </Box>
-                <Box sx={{ pr: '2rem' }}>
-                    {renderUserData()}
-                </Box>
+                <Box sx={{ pr: '2rem' }}>{renderUserData()}</Box>
             </Box>
             <CustomSnackbar
                 open={errorSnackbar}
