@@ -1,21 +1,17 @@
-import { Box, Button, Card, Typography } from '@mui/material';
-import Grid from '@mui/material/Grid';
-import { useEffect, useState, useRef } from 'react';
+import { Mic, Pause, PlayArrow, Stop } from '@mui/icons-material';
+import { Box, Button, Card, Grid2, Typography } from '@mui/material';
+import { JSX, useEffect, useRef, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import MicIcon from '@mui/icons-material/Mic';
-import StopIcon from '@mui/icons-material/Stop';
-import PlayArrowIcon from '@mui/icons-material/PlayArrow';
-import PauseIcon from '@mui/icons-material/Pause';
 
 import { useGetPatient } from '../../api/queries';
 
-import CustomSnackbar from 'components/custom_snackbar/custom_snackbar';
-import PatientCard from 'components/patient_card/patient_card';
-import SmartAppBar from 'components/smart_app_bar/smart_app_bar';
-import ConditionsSummary from 'components/conditions_summary/conditions_summary';
-import MedicationsSummary from 'components/medications_summary/medications_summary';
-import AllergiesSummary from 'components/allergies_summary/allergies_summary';
-import routes from 'routes';
+import AllergiesSummary from '@/components/allergies_summary/allergies_summary';
+import ConditionsSummary from '@/components/conditions_summary/conditions_summary';
+import CustomSnackbar from '@/components/custom_snackbar/custom_snackbar';
+import MedicationsSummary from '@/components/medications_summary/medications_summary';
+import PatientCard from '@/components/patient_card/patient_card';
+import SmartAppBar from '@/components/smart_app_bar/smart_app_bar';
+import routes from '@/routes';
 
 const PatientProfile = (): JSX.Element => {
     const [errorMessage, setErrorMessage] = useState<string>('');
@@ -47,7 +43,7 @@ const PatientProfile = (): JSX.Element => {
                     setPermissionStatus(newState);
                     setShowPermissionPrompt(newState === 'prompt');
                 });
-            } catch (err) {
+            } catch {
                 setPermissionStatus('error checking permissions');
                 setShowPermissionPrompt(true);
             }
@@ -61,17 +57,17 @@ const PatientProfile = (): JSX.Element => {
     const requestPermission = async (): Promise<void> => {
         try {
             // First try to get a temporary stream just to trigger the permission prompt
-            const stream = await navigator.mediaDevices.getUserMedia({ 
+            const stream = await navigator.mediaDevices.getUserMedia({
                 audio: true,
                 video: false
             });
             // Immediately stop the stream since we just wanted the permission
-            stream.getTracks().forEach(track => track.stop());
+            stream.getTracks().forEach((track) => track.stop());
             // Check the new permission state
             const permissions = await navigator.permissions.query({ name: 'microphone' as PermissionName });
             setPermissionStatus(permissions.state);
             setShowPermissionPrompt(permissions.state === 'prompt');
-            
+
             if (permissions.state === 'granted') {
                 setErrorMessage('Microphone access granted. You can now start recording.');
             }
@@ -93,7 +89,7 @@ const PatientProfile = (): JSX.Element => {
             }
             troubleshooting += `- Current permission status: ${permissionStatus}`;
         }
-        
+
         setErrorMessage(`${context}: ${err.name} - ${err.message}${troubleshooting}`);
     };
 
@@ -104,27 +100,30 @@ const PatientProfile = (): JSX.Element => {
             }
 
             // Debug info about permissions and environment
-            console.log('Checking permissions and environment...');
-            console.log('User Agent:', navigator.userAgent);
-            console.log('MediaDevices available:', !!navigator.mediaDevices);
-            
+            console.info('Checking permissions and environment...');
+            console.info('User Agent:', navigator.userAgent);
+            console.info('MediaDevices available:', !!navigator.mediaDevices);
+
             try {
                 const permissions = await navigator.permissions.query({ name: 'microphone' as PermissionName });
-                console.log('Microphone permission state:', permissions.state);
+                console.info('Microphone permission state:', permissions.state);
                 permissions.addEventListener('change', (e) => {
-                    console.log('Permission state changed:', (e.target as PermissionStatus).state);
+                    console.info('Permission state changed:', (e.target as PermissionStatus).state);
                 });
             } catch (permErr) {
-                console.log('Error querying permissions:', permErr);
+                console.info('Error querying permissions:', permErr);
             }
 
             // Log iframe context
-            console.log('Is in iframe:', window !== window.parent);
-            console.log('Iframe permissions policy:', 'featurePolicy' in document ? 
-                (document as any).featurePolicy?.allowsFeature('microphone') : 
-                'Feature Policy API not available');
+            console.info('Is in iframe:', window !== window.parent);
+            console.info(
+                'Iframe permissions policy:',
+                'featurePolicy' in document
+                    ? (document as any).featurePolicy?.allowsFeature('microphone')
+                    : 'Feature Policy API not available'
+            );
 
-            const stream = await navigator.mediaDevices.getUserMedia({ 
+            const stream = await navigator.mediaDevices.getUserMedia({
                 audio: {
                     echoCancellation: true,
                     noiseSuppression: true,
@@ -132,24 +131,27 @@ const PatientProfile = (): JSX.Element => {
                 },
                 video: false
             });
-            
-            console.log('Stream obtained successfully');
-            console.log('Audio tracks:', stream.getAudioTracks().map(track => ({
-                label: track.label,
-                enabled: track.enabled,
-                muted: track.muted,
-                readyState: track.readyState
-            })));
-            
+
+            console.info('Stream obtained successfully');
+            console.info(
+                'Audio tracks:',
+                stream.getAudioTracks().map((track) => ({
+                    label: track.label,
+                    enabled: track.enabled,
+                    muted: track.muted,
+                    readyState: track.readyState
+                }))
+            );
+
             const recorder = new MediaRecorder(stream, {
                 mimeType: MediaRecorder.isTypeSupported('audio/webm') ? 'audio/webm' : 'audio/ogg'
             });
-            console.log('MediaRecorder created with mimeType:', recorder.mimeType);
+            console.info('MediaRecorder created with mimeType:', recorder.mimeType);
 
             const chunks: BlobPart[] = [];
 
             recorder.ondataavailable = (e) => {
-                console.log('Data available event, chunk size:', e.data.size);
+                console.info('Data available event, chunk size:', e.data.size);
                 if (e.data.size > 0) {
                     chunks.push(e.data);
                 }
@@ -160,15 +162,15 @@ const PatientProfile = (): JSX.Element => {
             };
 
             recorder.onstop = () => {
-                console.log('Recording stopped, creating blob from chunks:', chunks.length);
+                console.info('Recording stopped, creating blob from chunks:', chunks.length);
                 const blob = new Blob(chunks, { type: recorder.mimeType });
-                console.log('Blob created, size:', blob.size);
+                console.info('Blob created, size:', blob.size);
                 const url = URL.createObjectURL(blob);
                 setAudioUrl(url);
             };
 
             recorder.start(1000); // Get data every second
-            console.log('Recording started');
+            console.info('Recording started');
             setMediaRecorder(recorder);
             setIsRecording(true);
 
@@ -192,7 +194,7 @@ const PatientProfile = (): JSX.Element => {
         try {
             if (mediaRecorder && mediaRecorder.state !== 'inactive') {
                 mediaRecorder.stop();
-                mediaRecorder.stream.getTracks().forEach(track => track.stop());
+                mediaRecorder.stream.getTracks().forEach((track) => track.stop());
                 setIsRecording(false);
                 setMediaRecorder(null);
             }
@@ -212,7 +214,7 @@ const PatientProfile = (): JSX.Element => {
             if (isPlaying) {
                 audioRef.current.pause();
             } else {
-                audioRef.current.play().catch(err => {
+                audioRef.current.play().catch((err) => {
                     handleError(err, 'Failed to play audio');
                 });
             }
@@ -223,11 +225,14 @@ const PatientProfile = (): JSX.Element => {
         }
     };
 
-    useEffect(() => () => {
-        if (audioUrl) {
-            URL.revokeObjectURL(audioUrl);
-        }
-    }, [audioUrl]);
+    useEffect(
+        () => () => {
+            if (audioUrl) {
+                URL.revokeObjectURL(audioUrl);
+            }
+        },
+        [audioUrl]
+    );
 
     const getPermissionColor = (status: string): string => {
         switch (status) {
@@ -254,8 +259,13 @@ const PatientProfile = (): JSX.Element => {
                     p: '2rem'
                 }}
             >
-                <Grid container spacing={2}>
-                    <Grid item xs={12} md={6}>
+                <Grid2 container spacing={2}>
+                    <Grid2
+                        size={{
+                            xs: 12,
+                            md: 6
+                        }}
+                    >
                         <Box
                             sx={{
                                 display: 'flex',
@@ -266,8 +276,13 @@ const PatientProfile = (): JSX.Element => {
                         >
                             <PatientCard patient={data} isLoading={isLoading} />
                         </Box>
-                    </Grid>
-                    <Grid item xs={12} md={6}>
+                    </Grid2>
+                    <Grid2
+                        size={{
+                            xs: 12,
+                            md: 6
+                        }}
+                    >
                         <Box
                             sx={{
                                 display: 'flex',
@@ -276,17 +291,27 @@ const PatientProfile = (): JSX.Element => {
                                 height: '100%'
                             }}
                         >
-                            <Grid container spacing={2}>
-                                <Grid item xs={12}>
+                            <Grid2 container spacing={2}>
+                                <Grid2 size={12}>
                                     <ConditionsSummary />
-                                </Grid>
-                                <Grid item xs={12} md={6}>
+                                </Grid2>
+                                <Grid2
+                                    size={{
+                                        xs: 12,
+                                        md: 6
+                                    }}
+                                >
                                     <MedicationsSummary />
-                                </Grid>
-                                <Grid item xs={12} md={6}>
+                                </Grid2>
+                                <Grid2
+                                    size={{
+                                        xs: 12,
+                                        md: 6
+                                    }}
+                                >
                                     <AllergiesSummary />
-                                </Grid>
-                                <Grid item xs={12}>
+                                </Grid2>
+                                <Grid2 size={12}>
                                     <Card
                                         sx={{
                                             p: 2,
@@ -300,13 +325,20 @@ const PatientProfile = (): JSX.Element => {
                                         <Typography variant="body2" color={getPermissionColor(permissionStatus)}>
                                             Microphone Permission: {permissionStatus}
                                         </Typography>
-                                        <Box sx={{ display: 'flex', gap: 2, flexDirection: 'column', alignItems: 'center' }}>
+                                        <Box
+                                            sx={{
+                                                display: 'flex',
+                                                gap: 2,
+                                                flexDirection: 'column',
+                                                alignItems: 'center'
+                                            }}
+                                        >
                                             {(showPermissionPrompt || permissionStatus === 'denied') && (
                                                 <Button
                                                     variant="outlined"
                                                     color="primary"
                                                     onClick={requestPermission}
-                                                    startIcon={<MicIcon />}
+                                                    startIcon={<Mic />}
                                                 >
                                                     Request Microphone Access
                                                 </Button>
@@ -314,21 +346,21 @@ const PatientProfile = (): JSX.Element => {
                                             <Box sx={{ display: 'flex', gap: 2 }}>
                                                 <Button
                                                     variant="contained"
-                                                    color={isRecording ? "error" : "primary"}
-                                                    startIcon={isRecording ? <StopIcon /> : <MicIcon />}
+                                                    color={isRecording ? 'error' : 'primary'}
+                                                    startIcon={isRecording ? <Stop /> : <Mic />}
                                                     onClick={isRecording ? stopRecording : startRecording}
                                                     disabled={permissionStatus !== 'granted'}
                                                 >
-                                                    {isRecording ? "Stop Recording" : "Start Recording"}
+                                                    {isRecording ? 'Stop Recording' : 'Start Recording'}
                                                 </Button>
                                                 {audioUrl && (
                                                     <Button
                                                         variant="contained"
                                                         color="secondary"
-                                                        startIcon={isPlaying ? <PauseIcon /> : <PlayArrowIcon />}
+                                                        startIcon={isPlaying ? <Pause /> : <PlayArrow />}
                                                         onClick={togglePlayback}
                                                     >
-                                                        {isPlaying ? "Pause" : "Play"}
+                                                        {isPlaying ? 'Pause' : 'Play'}
                                                     </Button>
                                                 )}
                                             </Box>
@@ -340,26 +372,18 @@ const PatientProfile = (): JSX.Element => {
                                                 onEnded={() => setIsPlaying(false)}
                                                 style={{ display: 'none' }}
                                             >
-                                                <track 
-                                                    kind="captions" 
-                                                    srcLang="en" 
-                                                    label="English captions" 
-                                                />
+                                                <track kind="captions" srcLang="en" label="English captions" />
                                             </audio>
                                         )}
                                     </Card>
-                                </Grid>
-                            </Grid>
-                            <Button
-                                variant='contained'
-                                sx={{ mt: '0.5rem' }}
-                                onClick={() => navigate(routes.goals)}
-                            >
+                                </Grid2>
+                            </Grid2>
+                            <Button variant="contained" sx={{ mt: '0.5rem' }} onClick={() => navigate(routes.goals)}>
                                 Goals
                             </Button>
                         </Box>
-                    </Grid>
-                </Grid>
+                    </Grid2>
+                </Grid2>
             </Box>
         </>
     );
