@@ -1,135 +1,135 @@
 import { createTheme } from '@mui/material/styles';
 
 const theme = createTheme({
-  palette: {
-    primary: {
-      main: '#2563EB', // Modern blue
-      light: '#60A5FA',
-      dark: '#1D4ED8',
-      contrastText: '#ffffff',
-    },
-    secondary: {
-      main: '#059669', // Modern green
-      light: '#34D399',
-      dark: '#047857',
-      contrastText: '#ffffff',
-    },
-    background: {
-      default: '#F8FAFC',
-      paper: '#FFFFFF',
-    },
-    text: {
-      primary: '#1E293B',
-      secondary: '#64748B',
-    },
-    grey: {
-      50: '#F8FAFC',
-      100: '#F1F5F9',
-      200: '#E2E8F0',
-      300: '#CBD5E1',
-      400: '#94A3B8',
-      500: '#64748B',
-      600: '#475569',
-      700: '#334155',
-      800: '#1E293B',
-      900: '#0F172A',
-    },
-  },
-  typography: {
-    fontFamily: '"Inter", "Roboto", "Helvetica", "Arial", sans-serif',
-    h4: {
-      fontWeight: 600,
-      letterSpacing: '-0.02em',
-      lineHeight: 1.2,
-    },
-    h5: {
-      fontWeight: 600,
-      letterSpacing: '-0.02em',
-      lineHeight: 1.2,
-    },
-    h6: {
-      fontWeight: 600,
-      letterSpacing: '-0.02em',
-      lineHeight: 1.2,
-    },
-    body1: {
-      lineHeight: 1.6,
-      letterSpacing: '0.01em',
-    },
-    body2: {
-      lineHeight: 1.6,
-      letterSpacing: '0.01em',
-    },
-  },
-  shape: {
-    borderRadius: 12,
-  },
-  components: {
-    MuiCard: {
-      styleOverrides: {
-        root: {
-          boxShadow: '0 1px 3px rgba(0,0,0,0.1), 0 1px 2px rgba(0,0,0,0.06)',
-          transition: 'all 0.2s ease-in-out',
-          '&:hover': {
-            boxShadow: '0 4px 6px rgba(0,0,0,0.1), 0 2px 4px rgba(0,0,0,0.06)',
-          },
+    palette: {
+        primary: {
+            main: '#2563EB', // Modern blue
+            light: '#60A5FA',
+            dark: '#1D4ED8',
+            contrastText: '#ffffff'
         },
-      },
-    },
-    MuiButton: {
-      styleOverrides: {
-        root: {
-          textTransform: 'none',
-          fontWeight: 500,
-          padding: '8px 16px',
-          borderRadius: 8,
+        secondary: {
+            main: '#059669', // Modern green
+            light: '#34D399',
+            dark: '#047857',
+            contrastText: '#ffffff'
         },
-        contained: {
-          boxShadow: 'none',
-          '&:hover': {
-            boxShadow: '0 2px 4px rgba(0,0,0,0.1)',
-          },
+        background: {
+            default: '#F8FAFC',
+            paper: '#FFFFFF'
         },
         text: {
-          '&:hover': {
-            backgroundColor: 'rgba(0,0,0,0.04)',
-          },
+            primary: '#1E293B',
+            secondary: '#64748B'
         },
-      },
+        grey: {
+            50: '#F8FAFC',
+            100: '#F1F5F9',
+            200: '#E2E8F0',
+            300: '#CBD5E1',
+            400: '#94A3B8',
+            500: '#64748B',
+            600: '#475569',
+            700: '#334155',
+            800: '#1E293B',
+            900: '#0F172A'
+        }
     },
-    MuiAppBar: {
-      styleOverrides: {
-        root: {
-          boxShadow: '0 1px 3px rgba(0,0,0,0.1), 0 1px 2px rgba(0,0,0,0.06)',
-          background: '#FFFFFF',
-          color: '#1E293B',
-          borderBottom: '1px solid #E2E8F0',
+    typography: {
+        fontFamily: '"Inter", "Roboto", "Helvetica", "Arial", sans-serif',
+        h4: {
+            fontWeight: 600,
+            letterSpacing: '-0.02em',
+            lineHeight: 1.2
         },
-      },
-    },
-    MuiCheckbox: {
-      styleOverrides: {
-        root: {
-          '&.Mui-checked': {
-            color: '#2563EB',
-          },
+        h5: {
+            fontWeight: 600,
+            letterSpacing: '-0.02em',
+            lineHeight: 1.2
         },
-      },
-    },
-    MuiPaginationItem: {
-      styleOverrides: {
-        root: {
-          '&.Mui-selected': {
-            backgroundColor: '#2563EB',
-            color: '#FFFFFF',
-            '&:hover': {
-              backgroundColor: '#1D4ED8',
-            },
-          },
+        h6: {
+            fontWeight: 600,
+            letterSpacing: '-0.02em',
+            lineHeight: 1.2
         },
-      },
+        body1: {
+            lineHeight: 1.6,
+            letterSpacing: '0.01em'
+        },
+        body2: {
+            lineHeight: 1.6,
+            letterSpacing: '0.01em'
+        }
     },
-  },
+    shape: {
+        borderRadius: 12
+    },
+    components: {
+        MuiCard: {
+            styleOverrides: {
+                root: {
+                    boxShadow: '0 1px 3px rgba(0,0,0,0.1), 0 1px 2px rgba(0,0,0,0.06)',
+                    transition: 'all 0.2s ease-in-out',
+                    '&:hover': {
+                        boxShadow: '0 4px 6px rgba(0,0,0,0.1), 0 2px 4px rgba(0,0,0,0.06)'
+                    }
+                }
+            }
+        },
+        MuiButton: {
+            styleOverrides: {
+                root: {
+                    textTransform: 'none',
+                    fontWeight: 500,
+                    padding: '8px 16px',
+                    borderRadius: 8
+                },
+                contained: {
+                    boxShadow: 'none',
+                    '&:hover': {
+                        boxShadow: '0 2px 4px rgba(0,0,0,0.1)'
+                    }
+                },
+                text: {
+                    '&:hover': {
+                        backgroundColor: 'rgba(0,0,0,0.04)'
+                    }
+                }
+            }
+        },
+        MuiAppBar: {
+            styleOverrides: {
+                root: {
+                    boxShadow: '0 1px 3px rgba(0,0,0,0.1), 0 1px 2px rgba(0,0,0,0.06)',
+                    background: '#FFFFFF',
+                    color: '#1E293B',
+                    borderBottom: '1px solid #E2E8F0'
+                }
+            }
+        },
+        MuiCheckbox: {
+            styleOverrides: {
+                root: {
+                    '&.Mui-checked': {
+                        color: '#2563EB'
+                    }
+                }
+            }
+        },
+        MuiPaginationItem: {
+            styleOverrides: {
+                root: {
+                    '&.Mui-selected': {
+                        backgroundColor: '#2563EB',
+                        color: '#FFFFFF',
+                        '&:hover': {
+                            backgroundColor: '#1D4ED8'
+                        }
+                    }
+                }
+            }
+        }
+    }
 });
 
-export default theme; 
+export default theme;
